+++ src/cuda/utils.jl
@@ -1,0 +1,2 @@
+
+Flux.rng_from_array(::CuArray) = CUDA.default_rng()
